fix: Adiciona validação na exclusão de status do item

+++ src/main/java/br/com/fatec/mogi/inventory_service/coreService/service/impl/StatusItemServiceImpl.java
@@ -2,7 +2,9 @@ package br.com.fatec.mogi.inventory_service.coreService.service.impl;
 
 import br.com.fatec.mogi.inventory_service.coreService.domain.exception.StatusItemJaCadastradoException;
 import br.com.fatec.mogi.inventory_service.coreService.domain.exception.StatusItemNaoEncontradoException;
+import br.com.fatec.mogi.inventory_service.coreService.domain.exception.StatusItemNaoPodeSerExcluidoException;
 import br.com.fatec.mogi.inventory_service.coreService.domain.model.StatusItem;
+import br.com.fatec.mogi.inventory_service.coreService.repository.ItemRepository;
 import br.com.fatec.mogi.inventory_service.coreService.repository.StatusItemRepository;
 import br.com.fatec.mogi.inventory_service.coreService.service.StatusItemService;
 import br.com.fatec.mogi.inventory_service.coreService.web.request.CadastrarStatusItemRequestDTO;
@@ -14,6 +16,8 @@ import org.springframework.stereotype.Service;
 public class StatusItemServiceImpl implements StatusItemService {
 
 	private final StatusItemRepository statusItemRepository;
+
+	private final ItemRepository itemRepository;
 
 	@Override
 	public void cadastrar(CadastrarStatusItemRequestDTO dto) {
@@ -27,6 +31,9 @@ public class StatusItemServiceImpl implements StatusItemService {
 	@Override
 	public void deletar(Long id) {
 		statusItemRepository.findById(id).orElseThrow(StatusItemNaoEncontradoException::new);
+		if (itemRepository.existsByStatusItemId(id)) {
+			throw new StatusItemNaoPodeSerExcluidoException();
+		}
 		statusItemRepository.deleteById(id);
 	}
